test-llm-container/requirements.txt
fastapi
uvicorn
httpx
azure-identity
azure-keyvault-secrets